+++ static/index.html
@@ -31,7 +31,7 @@
 					</table>
 				</div>
 				<div id="chatentry">
-					<input autofocus placeholder="Message..." id="input" maxlength="500">
+					<input autofocus placeholder="Message ou attaque avec '/attack <pokémon>'" id="input" maxlength="500">
 				</div>
 			</div>
 			<div id="userlist">
