+++ tools/antiflood_system.md
@@ -44,3 +44,12 @@ Cette technique, en plus d'ête ludique, ne charge que peut le serveur: en effet
 nécessitent déjà 100 envois (et le serveur s'en sort très bien). Ainsi; 100 messages d'un coup sur un unique client n'est pas un poids
 conséquent pour le serveur, mais plutot lourd pour le navigateur du flooder qui va devoir traiter un sacré paquets de messages d'un coup
 (sans parler du lecteur audio qui devra mixer 100 tracks wav d'un coup).
+
+## Todo
+
+ * MàJ des users pour avoir un historique des messages et un indicateur de bann
+ * MàJ du handler de message pour gérer un historique des messages de l'user
+ * MàJ du hangler de message pour déclencher un bann quand l'user a trop flooder
+ * MàJ du handler de message pour gérer les bannis shadowmute 
+ * MàJ des attaques pours permettre l'attaque des bannis
+ * Ajout des notifications et effets audios relatifs au bann
