+++ tools/combat.py
@@ -11,10 +11,11 @@ class CombatSimulator:
 
     def __init__(self):
         self._affected_users = []
+        self.atk_dice, self.atk_bonus, self.def_dice, self.def_bonus= None, None, None, None
 
     def _fumble(self, user):
         for effect in [get_random_effect() for i in range(4)]:
-            user.add_effect(effect)
+            user.state.add_effect(effect)
             self._affected_users.append((user, effect))
 
     def run_attack(self, attacker, defender, channel):
@@ -26,7 +27,7 @@ class CombatSimulator:
             for userid, user in channel.users.items():
                 if userid != attacker.user_id:
                     effect_obj = effect_type()
-                    user.add_effect(effect_obj)
+                    user.state.add_effect(effect_obj)
                     self._affected_users.append((user,effect_obj))
 
         elif self.atk_dice == 1 or self.def_dice == 100: # attack fumble
@@ -39,16 +40,16 @@ class CombatSimulator:
             randoum = random.randint(1,3)
             effect = get_random_effect()
             if randoum == 1: # bounceback
-                attacker.add_effect(effect)
+                attacker.state.add_effect(effect)
                 self._affected_users = [(attacker, effect)]
             elif randoum == 2:
                 random_other_user = random.choice(list(channel.users.values()))
-                random_other_user.add_effect(effect)
+                random_other_user.state.add_effect(effect)
                 self._affected_users = [(random_other_user, effect)]
 
         elif self.atk_dice + self.atk_bonus > self.def_dice + self.def_bonus:# regular atck pass
             effect = get_random_effect()
-            defender.add_effect(effect)
+            defender.state.add_effect(effect)
             self._affected_users = [(defender, effect)]
 
     @property
